requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

alembic>=0.9.6 # MIT
croniter>=0.3.4 # MIT License
cachetools>=2.0.0 # MIT License
dogpile.cache>=0.6.2 # BSD
eventlet>=0.26.0 # MIT
Jinja2>=2.10 # BSD License (3 clause)
jsonschema>=3.2.0 # MIT
keystonemiddleware>=4.18.0 # Apache-2.0
kombu!=4.0.2,>=4.6.1 # BSD
mistral-lib==3.0.0 # Apache-2.0
networkx>=2.3 # BSD
oslo.concurrency>=3.26.0 # Apache-2.0
oslo.config>=6.8.0 # Apache-2.0
oslo.context>=2.22.0 # Apache-2.0
oslo.db>=4.40.0 # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
oslo.messaging>=14.1.0 # Apache-2.0
oslo.middleware>=3.31.0 # Apache-2.0
oslo.policy>=3.6.0 # Apache-2.0
oslo.utils>=4.0.0 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.serialization>=2.21.1 # Apache-2.0
oslo.service>=2.1.0 # Apache-2.0
osprofiler>=1.4.0 # Apache-2.0
paramiko>=2.4.1 # LGPLv2.1+
pbr!=2.1.0,>=2.0.0 # Apache-2.0
pecan>=1.2.1 # BSD
PyJWT>=1.5 # MIT
PyYAML>=5.1 # MIT
requests>=2.18.0 # Apache-2.0
tenacity>=5.0.1 # Apache-2.0
SQLAlchemy>=1.2.5,<2.0.0 # MIT
stevedore>=1.20.0 # Apache-2.0
WSME>=0.8.0 # MIT
yaql>=1.1.3 # Apache 2.0 License
tooz>=1.58.0 # Apache-2.0
zake>=0.1.6 # Apache-2.0